feat: migrate email sending from Resend to Brevo
- Added @getbrevo/brevo dependency for email handling.
- Updated environment configuration to use BREVO_API_KEY instead of RESEND_API_KEY.
- Removed Resend configuration and helper functions.
- Created new Brevo configuration and helper functions for sending overdue emails.
- Updated job to utilize the new Brevo email sending functionality.
