--- src/utils/resend.helper.ts
@@ -1,28 +1,0 @@
-import { env } from "../config/env.config"
-import { resend } from "../config/resend.config";
-import { toBoliviaTime } from "./date.helper"
-
-export async function sendOverdueEmail(
-  targetMail: string,
-  userName: string,
-  gameTitle: string,
-  limitDate: Date) {
-  try {
-    const formatedLimitDate = toBoliviaTime(limitDate.toISOString());
-
-    await resend.emails.send({
-      from: `${env.SEND_USER} <${env.SEND_EMAIL_USER}>`,
-      to: targetMail,
-      subject: "Préstamo vencido - Devolución requerida",
-      html: `
-      <h2>Hola, ${userName}</h2>
-      <p>Tu préstamo del juego <strong>${gameTitle}</strong> venció el 
-      <strong>${formatedLimitDate}</strong> y aún no ha sido devuelto.</p>
-      <p>Por favor acércate a devolver el juego a la brevedad posible.</p>
-      <p><strong>Se aplicará una multa por cada día de retraso.</strong></p>
-    `
-    });
-  } catch (err) {
-    throw new Error(`Error al enviar correo electronico: ${(err as Error).message}`);
-  }
-}
